Moved predicates to main folder as 'comparison' is redundant
All predicates on non-boolean values will be about comparing it in some way, no reason to have a separate folder
