+++ src/main/java/com/minecraftabnormals/abnormals_core/core/api/conditions/config_predicates/GreaterThanPredicate.java
@@ -1,20 +1,18 @@
-package com.minecraftabnormals.abnormals_core.core.api.conditions.config_predicates.comparison;
+package com.minecraftabnormals.abnormals_core.core.api.conditions.config_predicates;
 
 import com.google.gson.JsonObject;
 import com.google.gson.JsonSyntaxException;
 import com.minecraftabnormals.abnormals_core.core.AbnormalsCore;
-import com.minecraftabnormals.abnormals_core.core.api.conditions.config_predicates.IConfigPredicate;
-import com.minecraftabnormals.abnormals_core.core.api.conditions.config_predicates.IConfigPredicateSerializer;
 import net.minecraft.util.ResourceLocation;
 import net.minecraftforge.common.ForgeConfigSpec;
 
 import java.math.BigDecimal;
 
-public class GreaterThanOrEqualPredicate implements IConfigPredicate {
-    private static final ResourceLocation ID = new ResourceLocation(AbnormalsCore.MODID, "greater_than_or_equal_to");
+public class GreaterThanPredicate implements IConfigPredicate {
+    private static final ResourceLocation ID = new ResourceLocation(AbnormalsCore.MODID, "greater_than");
     private final BigDecimal value;
 
-    public GreaterThanOrEqualPredicate(BigDecimal value) {
+    public GreaterThanPredicate(BigDecimal value) {
         this.value = value;
     }
 
@@ -27,27 +25,27 @@ public class GreaterThanOrEqualPredicate implements IConfigPredicate {
     public boolean test(ForgeConfigSpec.ConfigValue<?> toCompare) {
         try {
             BigDecimal number = new BigDecimal(toCompare.get().toString());
-            return number.compareTo(value) >= 0;
+            return number.compareTo(value) > 0;
         } catch (NumberFormatException e) {
             throw new IllegalArgumentException("Invalid config value type; must hold a Number");
         }
     }
 
-    public static class Serializer implements IConfigPredicateSerializer<GreaterThanOrEqualPredicate> {
-        private static final ResourceLocation ID = new ResourceLocation(AbnormalsCore.MODID, "greater_than_or_equal_to");
+    public static class Serializer implements IConfigPredicateSerializer<GreaterThanPredicate> {
+        private static final ResourceLocation ID = new ResourceLocation(AbnormalsCore.MODID, "greater_than");
 
         @Override
         public void write(JsonObject json, IConfigPredicate value) {
-            if (!(value instanceof GreaterThanOrEqualPredicate)) throw new IllegalArgumentException("Incompatible predicate type");
-            json.addProperty("value", ((GreaterThanOrEqualPredicate) value).value);
+            if (!(value instanceof GreaterThanPredicate)) throw new IllegalArgumentException("Incompatible predicate type");
+            json.addProperty("value", ((GreaterThanPredicate) value).value);
         }
 
         @Override
-        public GreaterThanOrEqualPredicate read(JsonObject json) {
+        public GreaterThanPredicate read(JsonObject json) {
             if (!json.has("value"))
                 throw new JsonSyntaxException("Missing 'value', expected to find a number");
             try {
-                return new GreaterThanOrEqualPredicate(json.get("value").getAsBigDecimal());
+                return new GreaterThanPredicate(json.get("value").getAsBigDecimal());
             } catch (NumberFormatException e) {
                 throw new JsonSyntaxException("'value' does not contain a number");
             }
